requirements.txt
python_dotenv
langchain
